+++ shelf_packages/shelf_routing_generator/example/body/body_parameters.g.dart
@@ -13,13 +13,13 @@ Router _$PathParametersControllerRouter(PathParametersController service) =>
       ..add('POST', '/integer', (Request request) async {
         return service.integer(
           request,
-          await $parseBodyAs(request, (data) => data! as int),
+          await $readBodyAs(request, (data) => data! as int),
         );
       })
       ..add('POST', '/fromJson', (Request request) async {
         return service.fromJson(
           request,
-          await $parseBodyAs(
+          await $readBodyAs(
             request,
             (data) => Decimal.fromJson(data! as String),
           ),
@@ -28,7 +28,7 @@ Router _$PathParametersControllerRouter(PathParametersController service) =>
       ..add('POST', '/listDouble', (Request request) async {
         return service.listDouble(
           request,
-          await $parseBodyAs(
+          await $readBodyAs(
             request,
             (data) => (data! as List<dynamic>).map((data) {
               return data! as double;
@@ -39,7 +39,7 @@ Router _$PathParametersControllerRouter(PathParametersController service) =>
       ..add('POST', '/listFromJson', (Request request) async {
         return service.listFromJson(
           request,
-          await $parseBodyAs(
+          await $readBodyAs(
             request,
             (data) => (data! as List<dynamic>).map((data) {
               return Decimal.fromJson(data! as String);
@@ -50,7 +50,7 @@ Router _$PathParametersControllerRouter(PathParametersController service) =>
       ..add('POST', '/mapNum', (Request request) async {
         return service.mapNum(
           request,
-          await $parseBodyAs(
+          await $readBodyAs(
             request,
             (data) => (data! as Map<String, dynamic>).map((k, data) {
               return MapEntry(k, data! as num);
@@ -61,7 +61,7 @@ Router _$PathParametersControllerRouter(PathParametersController service) =>
       ..add('POST', '/mapFromJson', (Request request) async {
         return service.mapFromJson(
           request,
-          await $parseBodyAs(
+          await $readBodyAs(
             request,
             (data) => (data! as Map<String, dynamic>).map((k, data) {
               return MapEntry(k, Decimal.fromJson(data! as String));