+++ shelf_packages/shelf_routing_generator/lib/src/route_handler.dart
@@ -1,39 +1,83 @@
 import 'package:analyzer/dart/element/element.dart';
 import 'package:analyzer/dart/element/type.dart';
+import 'package:http_methods/http_methods.dart';
+// ignore: implementation_imports
+import 'package:shelf_router/src/router_entry.dart' show RouterEntry;
 import 'package:shelf_routing_generator/src/route_header_handler.dart';
 import 'package:shelf_routing_generator/src/utils.dart';
 import 'package:source_gen/source_gen.dart';
 
 enum RouteReturnsType { response, json, nothing }
 
-sealed class RouteHandlerBase {
-  static List<RouteHandlerBase> fromClass(ClassElement element) {
+class _Route {
+  final String verb;
+  final String path;
+
+  _Route(this.verb, this.path);
+
+  static _Route? from(ExecutableElement element) {
+    final annotation = ConstantReader(routeChecker.firstAnnotationOf(element));
+    if (annotation.isNull) return null;
+
+    return _Route(annotation.read('verb').stringValue, annotation.read('route').stringValue);
+  }
+}
+
+sealed class RouteHandler {
+  final ExecutableElement element;
+  final String path;
+
+  RouteHandler({required this.element, required this.path});
+
+  static List<RouteHandler> from(ClassElement element, {bool strict = true}) {
     return [
-      ...element.methods.map(RouteHandler.from).nonNulls,
+      ...element.methods.map((element) => HttpRouteHandler.from(element, strict: strict)).nonNulls,
       ...element.accessors.map(MountRouteHandler.from).nonNulls,
     ];
   }
 }
 
-class MountRouteHandler extends RouteHandlerBase {
-  final ExecutableElement element;
+class MountRouteHandler extends RouteHandler {
   final bool isRouterMixin;
-  final String path;
 
-  MountRouteHandler({required this.element, required this.isRouterMixin, required this.path});
+  MountRouteHandler({required super.element, required super.path, required this.isRouterMixin});
 
   static MountRouteHandler? from(ExecutableElement element) {
-    final annotation = ConstantReader(routeChecker.firstAnnotationOf(element));
-    if (annotation.isNull) return null;
+    final route = _Route.from(element);
+    if (route == null) return null;
 
-    final verb = annotation.read('verb').stringValue;
-    final route = annotation.read('route').stringValue;
-    if (!RegExp(r'^\/.*').hasMatch(route)) {
-      throw InvalidGenerationSourceError('"route" field must begin with "/".', element: element);
-    }
-    if (verb != r'$mount') {
+    if (element.isStatic) {
       throw InvalidGenerationSourceError(
-        'the field must be use @Route.mount(...) or @Route.all(...) annotation.',
+        'The shelf_router.Route annotation cannot be used on static members',
+        element: element,
+      );
+    }
+    if (element.kind != ElementKind.GETTER) {
+      throw InvalidGenerationSourceError(
+        'The shelf_router.Route.mount annotation can only be used on a '
+        'getter that returns shelf_router.Router',
+        element: element,
+      );
+    }
+    if (!route.path.startsWith('/')) {
+      throw InvalidGenerationSourceError(
+        'The prefix "${route.path}" in shelf_router.Route.mount(prefix) '
+        'annotation must begin with a slash',
+        element: element,
+      );
+    }
+    if (route.path.contains('<')) {
+      throw InvalidGenerationSourceError(
+        'The prefix "${route.path}" in shelf_router.Route.mount(prefix) '
+        'annotation cannot contain <',
+        element: element,
+      );
+    }
+
+    if (route.verb != r'$mount') {
+      throw InvalidGenerationSourceError(
+        'Only the shelf_router.Route.mount annotation can only be used on a '
+        'getter, and only if it returns a shelf_router.Router or shelf_routing.RouterMixin',
         element: element,
       );
     }
@@ -46,45 +90,69 @@ class MountRouteHandler extends RouteHandlerBase {
       isRouterMixin = true;
     } else {
       throw InvalidGenerationSourceError(
-        'the getter must be a returns a "Router" type or a class with "router" getter to returns a "Router" type.',
+        'The getter must returns a shelf_router.Router or shelf_routing.RouterMixin',
         element: element,
       );
     }
 
-    return MountRouteHandler(element: element, isRouterMixin: isRouterMixin, path: route);
+    return MountRouteHandler(element: element, isRouterMixin: isRouterMixin, path: route.path);
   }
 }
 
-class RouteHandler extends RouteHandlerBase {
-  final ExecutableElement element;
-  final String method;
-  final String path;
-
+class HttpRouteHandler extends RouteHandler {
+  final String verb;
   final ParameterElement? bodyParameter;
   final bool hasRequest;
   final List<ParameterElement> pathParameters;
   final List<RouteHeaderHandler> headers;
-  // final List<ParameterElement> headerParameters;
   final List<ParameterElement> queryParameters;
   final RouteReturnsType returns;
 
-  static RouteHandler? from(MethodElement element) {
-    final annotation = ConstantReader(routeChecker.firstAnnotationOf(element));
-    if (annotation.isNull) return null;
+  static HttpRouteHandler? from(MethodElement element, {bool strict = true}) {
+    final route = _Route.from(element);
+    if (route == null) return null;
 
-    final verb = annotation.read('verb').stringValue;
-    final route = annotation.read('route').stringValue;
-    if (!RegExp(r'^\/.*').hasMatch(route)) {
-      throw InvalidGenerationSourceError('"route" field must begin with "/".', element: element);
+    if (element.isStatic) {
+      throw InvalidGenerationSourceError(
+        'The shelf_router.Route annotation cannot be used on static members',
+        element: element,
+      );
+    }
+    if (!isHttpMethod(route.verb)) {
+      throw InvalidGenerationSourceError(
+        'The verb "${route.verb}" used in shelf_router.Route annotation must be '
+        'a valid HTTP method',
+        element: element,
+      );
+    }
+    if (element.kind == ElementKind.GETTER) {
+      throw InvalidGenerationSourceError(
+        'Only the shelf_router.Route.mount annotation can only be used on a '
+        'getter, and only if it returns a shelf_router.Router',
+        element: element,
+      );
+    }
+    if (element.kind != ElementKind.METHOD) {
+      throw InvalidGenerationSourceError(
+        'The shelf_router.Route annotation can only be used on request handling methods',
+        element: element,
+      );
+    }
+    List<String> pathParams;
+    try {
+      pathParams = RouterEntry(route.verb, route.path, () => null).params;
+      // ignore: avoid_catching_errors
+    } on ArgumentError catch (e) {
+      throw InvalidGenerationSourceError(e.toString(), element: element);
     }
 
-    final pathParams = RegExp(r'\/<(\w+)>').allMatches(route).map((e) => e.group(1)!).toList();
     final parametersIterator = element.parameters.where((e) => e.isPositional).iterator;
 
     if (!parametersIterator.moveNext() ||
         !requestChecker.isAssignableFromType(parametersIterator.current.type)) {
       throw InvalidGenerationSourceError(
-        'need first parameter as "Request request".',
+        'The shelf_router.Route annotation can only be used on shelf request '
+        'handlers accept a shelf.Request parameter as first parameter',
         element: element,
       );
     }
@@ -93,7 +161,9 @@ class RouteHandler extends RouteHandlerBase {
     for (final pathParam in pathParams) {
       if (!parametersIterator.moveNext()) {
         throw InvalidGenerationSourceError(
-          'not has "$pathParam" path parameter.',
+          'The shelf_router.Route annotation can only be used on shelf '
+          'request handlers accept a shelf.Request parameter and all parameters in the route, '
+          'the "$pathParam" is missing',
           element: element,
         );
       }
@@ -101,7 +171,17 @@ class RouteHandler extends RouteHandlerBase {
 
       if (pathParam != parameter.name) {
         throw InvalidGenerationSourceError(
-          'has name different to path param "$pathParam".',
+          'The shelf_router.Route annotation can only be used on shelf '
+          'request handlers accept a shelf.Request parameter and and all parameters in the route, '
+          'the "${parameter.name}" parameter should be named "$pathParam"',
+          element: parameter,
+        );
+      }
+      if (parameter.isOptional) {
+        throw InvalidGenerationSourceError(
+          'The shelf_router.Route annotation can only be used on shelf '
+          'request handlers accept a shelf.Request parameter and all parameters in the route, '
+          'optional positional parameters are not permitted',
           element: parameter,
         );
       }
@@ -109,8 +189,8 @@ class RouteHandler extends RouteHandlerBase {
     }
 
     ParameterElement? bodyParameter;
-    if (parametersIterator.moveNext()) {
-      if (verb == 'GET') {
+    if (strict && parametersIterator.moveNext()) {
+      if (route.verb == 'GET') {
         throw InvalidGenerationSourceError('"GET" endpoint cannot have a body.', element: element);
       }
       final parameter = parametersIterator.current;
@@ -135,7 +215,7 @@ class RouteHandler extends RouteHandlerBase {
       }
 
       if (!check(parameter.type)) {
-        final parameterTypeName = parameter.type.getDisplayString(withNullability: false);
+        final parameterTypeName = parameter.type.getDisplayString();
         throw InvalidGenerationSourceError(
           'invalid body parameter type.\n'
           'You can use primitive types: bool, int, double, num or String or List, Map with primitive values.\n'
@@ -147,17 +227,35 @@ class RouteHandler extends RouteHandlerBase {
     }
 
     if (parametersIterator.moveNext()) {
-      throw InvalidGenerationSourceError('has many parameters.', element: element);
+      throw InvalidGenerationSourceError(
+        'The shelf_router.Route annotation can only be used on shelf '
+        'request handlers accept a shelf.Request parameter and all parameters in the route, '
+        'remote "${parametersIterator.current}" parameter from the method',
+        element: element,
+      );
     }
 
     // final headerParameters =
     //     element.parameters.where((e) => e.isNamed && _headerChecker.hasAnnotationOf(e)).toList();
-    final queryParameters = element.parameters.where((e) => e.isNamed).toList();
+    var queryParameters = <ParameterElement>[];
+    if (strict) {
+      queryParameters = element.parameters.where((e) => e.isNamed).toList();
+    } else {
+      for (final parameter in element.parameters) {
+        if (parameter.isNamed) {
+          throw InvalidGenerationSourceError(
+            'The shelf_router.Route annotation can only be used on shelf '
+            'request handlers accept a shelf.Request parameter and all parameters in the route, '
+            'optional positional parameters are not permitted',
+            element: parameter,
+          );
+        }
+      }
+    }
 
-    return RouteHandler._(
-      // routable: routable,
-      method: verb,
-      path: route,
+    return HttpRouteHandler._(
+      verb: route.verb,
+      path: route.path,
       element: element,
       bodyParameter: bodyParameter,
       hasRequest: true,
@@ -188,18 +286,17 @@ class RouteHandler extends RouteHandlerBase {
       return RouteReturnsType.json;
     }
 
-    final typeName = type.getDisplayString(withNullability: false);
+    final typeName = type.getDisplayString();
     throw InvalidGenerationSourceError(
       'Please update $typeName with valid returns json value.',
       element: type.element,
     );
   }
 
-  RouteHandler._({
-    // required this.routable,
-    required this.element,
-    required this.method,
-    required this.path,
+  HttpRouteHandler._({
+    required super.element,
+    required this.verb,
+    required super.path,
     required this.bodyParameter,
     required this.hasRequest,
     required this.pathParameters,