+++ shelf_packages/open_api_client_generator/CHANGELOG.md
@@ -1,3 +1,5 @@
+## 1.0.1
+- fix: fixed security field in open api operation
 
 ## 1.0.0
 - feat: First release of open_api_client_generator