+++ shelf_packages/open_api_specification/CHANGELOG.md
@@ -1,3 +1,5 @@
+## 2.0.0
+- fix!: fixed security field in open api operation
 
 ## 1.0.0
 - feat: First release of open_api_spec