+++ shelf_packages/shelf_open_api/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 3.0.0
+- feat!: routing has been left to the shelf_router and shelf_routing packages so the Routes annotation has been removed
+- feat!: now to generate the specs you will need to annotate a class with @OpenApi()
+
 ## 2.0.0
 - feat: The OpenApi for the entire project are generated directly in the public/open_api.yaml file